+++ app/controllers/authors_controller.rb
@@ -1,6 +1,7 @@
 class AuthorsController < ApplicationController
+
   before_action :authenticate_author!
 
   layout 'authors'
-  
+
 end
