+++ app/models/element.rb
@@ -1,3 +1,4 @@
 class Element < ApplicationRecord
   belongs_to :article
+  validates :element_type, inclusion: { in: ['paragraph', 'image', 'video'] }
 end
