+++ config/routes.rb
@@ -1,10 +1,11 @@
 Rails.application.routes.draw do
-  resources :elements
   root to: "staticpages#home"
   devise_for :authors
   # resources :articles
 
   scope module: 'authors' do
-    resources :articles
+    resources :articles do
+      resources :elements
+    end
   end
 end
